+++ pydid/drdid/utils.py
@@ -1,0 +1,47 @@
+"""Utility functions for the DRDiD package."""
+
+import warnings
+
+import numpy as np
+
+
+def _validate_inputs(arrays_dict, x, n_bootstrap, trim_level, check_intercept=False):
+    """Validate inputs for bootstrap functions."""
+    for name, arr in arrays_dict.items():
+        if not isinstance(arr, np.ndarray):
+            raise TypeError(f"{name} must be a NumPy array.")
+
+    if not isinstance(x, np.ndarray):
+        raise TypeError("x must be a NumPy array.")
+
+    for name, arr in arrays_dict.items():
+        if arr.ndim != 1:
+            raise ValueError(f"{name} must be 1-dimensional.")
+
+    if x.ndim != 2:
+        raise ValueError("x must be a 2-dimensional array.")
+
+    first_array = next(iter(arrays_dict.values()))
+    n_units = first_array.shape[0]
+
+    for name, arr in arrays_dict.items():
+        if arr.shape[0] != n_units:
+            raise ValueError("All arrays must have the same number of observations.")
+
+    if x.shape[0] != n_units:
+        raise ValueError("All arrays must have the same number of observations.")
+
+    if not isinstance(n_bootstrap, int) or n_bootstrap <= 0:
+        raise ValueError("n_bootstrap must be a positive integer.")
+
+    if not 0 < trim_level < 1:
+        raise ValueError("trim_level must be between 0 and 1.")
+
+    if check_intercept and not np.all(x[:, 0] == 1.0):
+        warnings.warn(
+            "The first column of the covariate matrix 'x' does not appear to be an intercept (all ones). "
+            "IPT propensity score estimation typically requires an intercept.",
+            UserWarning,
+        )
+
+    return n_units
