+++ pydid/utils.py
@@ -311,284 +311,6 @@ def preprocess_drdid(
     return output
 
 
-def preprocess_synth(
-    data: pd.DataFrame,
-    y_col: str,
-    time_col: str,
-    id_col: str,
-    treat_col: str,
-    treatment_period: Any,
-    covariates_formula: str | None = None,
-    l_outcome_periods: list[Any] | None = None,
-    weights_col: str | None = None,
-    normalized: bool = True,
-    post_periods_of_interest: list[Any] | None = None,
-) -> dict[str, Any]:
-    """Pre-processes data for synthetic control estimators.
-
-    Validates inputs, identifies treated/control units, balances panel data
-    for relevant pre- and post-treatment periods, processes covariates (patsy)
-    and lagged outcomes, normalizes weights, and structures output matrices
-    (Y_pre, Y_post, Z) for treated and control groups.
-
-    Parameters
-    ----------
-    data : pd.DataFrame
-        Input data. Must be panel data.
-    y_col : str
-        Outcome variable column name.
-    time_col : str
-        Time period indicator column name.
-    id_col : str
-        Unit identifier column name.
-    treat_col : str
-        Treatment indicator column name (0 for control, 1 for treated).
-        Assumed to be time-invariant for a unit.
-    treatment_period : Any
-        The time period when treatment begins for the 'treated' group.
-    covariates_formula : str or None, default None
-        Patsy formula for covariates (e.g., "~ x1 + x2"). Covariates are
-        taken from the last pre-treatment period or assumed time-invariant.
-    l_outcome_periods : list[Any] or None, default None
-        List of specific pre-treatment time periods whose outcome (y_col)
-        values should be used as predictors.
-    weights_col : str or None, default None
-        Observation weights column name. Assumed time-invariant per unit.
-    normalized : bool, default True
-        If True, unit weights are normalized to mean 1.
-    post_periods_of_interest : list[Any] or None, default None
-        Specific post-treatment periods for which outcomes should be reported.
-        If None, only `treatment_period` is used.
-
-    Returns
-    -------
-    dict[str, Any]
-        Dictionary with processed data:
-
-        - Y_pre_treat, Y_post_treat, Y_pre_control, Y_post_control: Outcome matrices.
-        - Z_treat, Z_control: Predictor matrices.
-        - weights_treat, weights_control: Unit weights.
-        - n_treated, n_control, n_pre_periods, n_post_periods: Counts.
-        - pre_periods, post_periods: Lists of period values.
-        - predictor_names: List of predictor names for Z matrices.
-    """
-    if not isinstance(data, pd.DataFrame):
-        warnings.warn("Input data is not a pandas DataFrame; converting...", UserWarning)
-        df = pd.DataFrame(data)
-    else:
-        df = data.copy()
-
-    base_required_cols = [y_col, time_col, id_col, treat_col]
-    if weights_col:
-        base_required_cols.append(weights_col)
-    missing_cols = [col for col in base_required_cols if col not in df.columns]
-    if missing_cols:
-        raise ValueError(f"Missing required columns: {', '.join(missing_cols)}")
-
-    for col in [y_col, time_col, id_col, treat_col]:
-        if not pd.api.types.is_numeric_dtype(df[col]) and col != id_col:
-            if pd.api.types.is_numeric_dtype(df[col]):
-                pass
-            elif col == time_col:
-                try:
-                    df[col] = pd.to_numeric(df[col])
-                    warnings.warn(f"Column '{col}' was not numeric; converted successfully.", UserWarning)
-                except (ValueError, TypeError) as exc:
-                    if not np.issubdtype(df[col].dtype, np.datetime64):
-                        raise TypeError(f"Column '{col}' must be numeric or datetime. Could not convert.") from exc
-            else:
-                try:
-                    df[col] = pd.to_numeric(df[col])
-                    warnings.warn(f"Column '{col}' was not numeric; converted successfully.", UserWarning)
-                except (ValueError, TypeError) as exc:
-                    raise TypeError(f"Column '{col}' must be numeric. Could not convert.") from exc
-
-    all_time_periods = sorted(df[time_col].unique())
-    if treatment_period not in all_time_periods:
-        raise ValueError(f"treatment_period ({treatment_period}) not found in time column '{time_col}'.")
-    if treatment_period == all_time_periods[0]:
-        raise ValueError("treatment_period cannot be the first period. Need pre-treatment periods.")
-
-    pre_periods = [t for t in all_time_periods if t < treatment_period]
-    if not pre_periods:
-        raise ValueError("No pre-treatment periods found before treatment_period.")
-
-    if l_outcome_periods:
-        if not all(p in pre_periods for p in l_outcome_periods):
-            raise ValueError("All l_outcome_periods must be in the pre-treatment period.")
-
-    default_post_periods = [treatment_period]
-    if post_periods_of_interest is None:
-        post_periods_of_interest = default_post_periods
-    else:
-        if not all(p >= treatment_period for p in post_periods_of_interest):
-            raise ValueError("All post_periods_of_interest must be >= treatment_period.")
-        if not all(p in all_time_periods for p in post_periods_of_interest):
-            raise ValueError("Not all post_periods_of_interest found in time column.")
-    post_periods_of_interest = sorted(list(set(post_periods_of_interest)))
-
-    relevant_periods = sorted(list(set(pre_periods + post_periods_of_interest)))
-    _check_treatment_uniqueness(df, id_col, treat_col)
-
-    treated_ids = df[df[treat_col] == 1][id_col].unique()
-    control_ids = df[df[treat_col] == 0][id_col].unique()
-
-    if len(treated_ids) == 0:
-        raise ValueError("No treated units found (treat_col == 1).")
-    if len(control_ids) == 0:
-        raise ValueError("No control units found (treat_col == 0) to form a donor pool.")
-
-    df_subset = df[df[id_col].isin(np.concatenate([treated_ids, control_ids]))]
-    df_balanced = _filter_and_balance_for_periods(df_subset, id_col, time_col, relevant_periods)
-    if df_balanced.empty:
-        raise ValueError("DataFrame empty after balancing for relevant periods. Check data coverage.")
-
-    final_all_ids = df_balanced[id_col].unique()
-    final_treated_ids = sorted(list(set(treated_ids).intersection(final_all_ids)))
-    final_control_ids = sorted(list(set(control_ids).intersection(final_all_ids)))
-
-    if not final_treated_ids:
-        raise ValueError("No treated units remain after balancing panel for all relevant periods.")
-    if not final_control_ids:
-        raise ValueError("No control units remain after balancing panel for all relevant periods.")
-
-    df_balanced = df_balanced.sort_values(by=[id_col, time_col])
-
-    na_check_cols = [y_col]
-    if weights_col:
-        na_check_cols.append(weights_col)
-
-    ids_with_na = df_balanced[df_balanced[time_col].isin(relevant_periods)][na_check_cols + [id_col]]
-    ids_with_na = ids_with_na.groupby(id_col).filter(lambda x: x[na_check_cols].isna().any().any())
-    ids_with_na = ids_with_na[id_col].unique()
-
-    if len(ids_with_na) > 0:
-        msg = f"Dropping {len(ids_with_na)} units due to NA in '{y_col}' or '{weights_col}' within relevant periods."
-        warnings.warn(msg, UserWarning)
-        df_balanced = df_balanced[~df_balanced[id_col].isin(ids_with_na)]
-        final_treated_ids = sorted(list(set(final_treated_ids) - set(ids_with_na)))
-        final_control_ids = sorted(list(set(final_control_ids) - set(ids_with_na)))
-
-    if not final_treated_ids:
-        raise ValueError("No treated units remain after NA handling.")
-    if not final_control_ids:
-        raise ValueError("No control units remain after NA handling.")
-
-    if weights_col:
-        df_weights = df_balanced[df_balanced[time_col] == relevant_periods[0]][[id_col, weights_col]].copy()
-        df_weights = df_weights.rename(columns={weights_col: "_unit_weights"})
-        if not pd.api.types.is_numeric_dtype(df_weights["_unit_weights"]):
-            raise TypeError(f"Weights column '{weights_col}' must be numeric.")
-        if (df_weights["_unit_weights"] < 0).any():
-            warnings.warn("Some weights are negative. Ensure this is intended.", UserWarning)
-    else:
-        df_weights = pd.DataFrame({id_col: df_balanced[id_col].unique(), "_unit_weights": 1.0})
-
-    weights_map = df_weights.set_index(id_col)["_unit_weights"]
-    weights_treat_arr = weights_map.loc[final_treated_ids].values.astype(float)
-    weights_control_arr = weights_map.loc[final_control_ids].values.astype(float)
-
-    if normalized:
-        overall_mean_weight = weights_map.loc[np.concatenate([final_treated_ids, final_control_ids])].mean()
-        if overall_mean_weight > 0:
-            weights_treat_arr /= overall_mean_weight
-            weights_control_arr /= overall_mean_weight
-        else:
-            warnings.warn("Mean of weights is zero or negative. Cannot normalize.", UserWarning)
-
-    df_pivot_y = df_balanced.pivot_table(index=id_col, columns=time_col, values=y_col)
-
-    y_pre_treat_df = df_pivot_y.loc[final_treated_ids, pre_periods]
-    y_post_treat_df = df_pivot_y.loc[final_treated_ids, post_periods_of_interest]
-    y_pre_control_df = df_pivot_y.loc[final_control_ids, pre_periods]
-    y_post_control_df = df_pivot_y.loc[final_control_ids, post_periods_of_interest]
-
-    predictor_names = []
-    z_parts_treat = []
-    z_parts_control = []
-
-    if l_outcome_periods:
-        lagged_outcomes_df = df_pivot_y[l_outcome_periods]
-        z_parts_treat.append(lagged_outcomes_df.loc[final_treated_ids].T.values)
-        z_parts_control.append(lagged_outcomes_df.loc[final_control_ids].T.values)
-        predictor_names.extend([f"{y_col}_lag_{p}" for p in l_outcome_periods])
-
-    generated_cov_names = []
-
-    if covariates_formula:
-        last_pre_period = pre_periods[-1]
-        df_for_formulaic = df_balanced[
-            (df_balanced[time_col] == last_pre_period)
-            & (df_balanced[id_col].isin(np.concatenate([final_treated_ids, final_control_ids])))
-        ].set_index(id_col)
-
-        try:
-            model_matrix_full = fml.model_matrix(covariates_formula, df_for_formulaic, output="pandas")
-            generated_cov_names = [col for col in model_matrix_full.columns if col != "Intercept"]
-
-            if generated_cov_names:
-                cov_matrix_treat = model_matrix_full.loc[final_treated_ids, generated_cov_names].T.values
-                cov_matrix_control = model_matrix_full.loc[final_control_ids, generated_cov_names].T.values
-                z_parts_treat.append(cov_matrix_treat)
-                z_parts_control.append(cov_matrix_control)
-                predictor_names.extend(generated_cov_names)
-        except Exception as e:
-            raise ValueError(
-                f"Error processing covariates_formula '{covariates_formula}' for synth with formulaic: {e}"
-            ) from e
-
-    z_treat = np.vstack(z_parts_treat) if z_parts_treat else np.empty((0, len(final_treated_ids)))
-    z_control = np.vstack(z_parts_control) if z_parts_control else np.empty((0, len(final_control_ids)))
-
-    if z_treat.shape[0] == 0:
-        msg = "No predictors specified (l_outcome_periods or covariates_formula). "
-        msg += "Z matrices will be empty."
-        warnings.warn(msg, UserWarning)
-
-    if z_control.shape[0] > 0 and z_control.shape[1] > 0:
-        if z_control.shape[0] < z_control.shape[1]:
-            matrix_for_rank_check = z_control.T
-            if matrix_for_rank_check.shape[0] > 0 and matrix_for_rank_check.shape[1] > 0:
-                _, r = np.linalg.qr(matrix_for_rank_check)
-                diag_r = np.abs(np.diag(r))
-                tol = diag_r.max() * 1e-7 if diag_r.size > 0 else 1e-7
-                rank = np.sum(diag_r > tol)
-                num_predictors = matrix_for_rank_check.shape[1]
-                if rank < num_predictors:
-                    msg = "Potential collinearity detected among predictors for control units. "
-                    msg += f"Rank ({rank}) is less than number of predictors ({num_predictors}). "
-                    msg += "Synthetic control weights may be unstable or not unique."
-                    warnings.warn(msg, UserWarning)
-        else:
-            msg = f"Number of predictors ({z_control.shape[0]}) >= "
-            msg += f"number of control units ({z_control.shape[1]}). "
-            msg += "This can lead to issues in synthetic control estimation."
-            warnings.warn(msg, UserWarning)
-
-    output = {
-        "Y_pre_treat": y_pre_treat_df.T.values,
-        "Y_post_treat": y_post_treat_df.T.values,
-        "Y_pre_control": y_pre_control_df.T.values,
-        "Y_post_control": y_post_control_df.T.values,
-        "Z_treat": z_treat,
-        "Z_control": z_control,
-        "weights_treat": weights_treat_arr,
-        "weights_control": weights_control_arr,
-        "n_treated": len(final_treated_ids),
-        "n_control": len(final_control_ids),
-        "n_pre_periods": len(pre_periods),
-        "n_post_periods": len(post_periods_of_interest),
-        "pre_periods": pre_periods,
-        "post_periods_of_interest": post_periods_of_interest,
-        "treatment_period": treatment_period,
-        "predictor_names": predictor_names,
-        "treated_ids": final_treated_ids,
-        "control_ids": final_control_ids,
-        "normalized_weights": normalized,
-    }
-    return output
-
-
 def _check_treatment_uniqueness(df: pd.DataFrame, id_col: str, treat_col: str) -> None:
     """Check if treatment status is unique for each ID in panel data."""
     treat_counts = df.groupby(id_col)[treat_col].nunique()
@@ -611,22 +333,3 @@ def _make_balanced_panel(df: pd.DataFrame, id_col: str, time_col: str) -> pd.Dat
             UserWarning,
         )
     return df[df[id_col].isin(ids_to_keep)].copy()
-
-
-def _filter_and_balance_for_periods(
-    df: pd.DataFrame, id_col: str, time_col: str, required_periods: list[Any]
-) -> pd.DataFrame:
-    """Filter dataframe for required_periods and keeps only units observed in all of them."""
-    df_filtered = df[df[time_col].isin(required_periods)].copy()
-    n_required_periods = len(set(required_periods))
-
-    obs_counts = df_filtered.groupby(id_col)[time_col].nunique()
-    ids_to_keep = obs_counts[obs_counts == n_required_periods].index
-
-    if len(ids_to_keep) < df_filtered[id_col].nunique():
-        msg = (
-            "Balancing panel for specified periods. Dropping units not observed in all "
-            f"{n_required_periods} required periods: {sorted(list(set(required_periods)))}."
-        )
-        warnings.warn(msg, UserWarning)
-    return df_filtered[df_filtered[id_col].isin(ids_to_keep)].copy()
