+++ tests/dgp.py
@@ -331,435 +331,6 @@ class DiD(BaseDGP):
         }
 
 
-class SyntheticControl(BaseDGP):
-    """Data generating process for synthetic control designs."""
-
-    def __init__(
-        self,
-        n_treated_units=1,
-        n_control_units=20,
-        n_time_pre=10,
-        n_time_post=5,
-        n_features=3,
-        random_seed=42,
-    ):
-        """Initialize the Synthetic Controls data generating process.
-
-        Parameters
-        ----------
-        n_treated_units : int, default=1
-            Number of units that receive treatment.
-        n_control_units : int, default=20
-            Number of units in the control pool.
-        n_time_pre : int, default=10
-            Number of time periods before treatment.
-        n_time_post : int, default=5
-            Number of time periods after treatment starts.
-        n_features : int, default=3
-            Number of covariates to generate.
-        random_seed : int, default=42
-            Random seed for reproducibility.
-        """
-        super().__init__(random_seed=random_seed)
-        self.n_treated_units = n_treated_units
-        self.n_control_units = n_control_units
-        self.n_time_pre = n_time_pre
-        self.n_time_post = n_time_post
-        self.n_features = n_features
-
-        if self.n_treated_units <= 0:
-            raise ValueError("n_treated_units must be positive.")
-        if self.n_control_units <= 0:
-            raise ValueError("n_control_units must be positive.")
-        if self.n_time_pre <= 0:
-            raise ValueError("n_time_pre must be positive.")
-        if self.n_time_post <= 0:
-            raise ValueError("n_time_post must be positive.")
-
-    def generate_data(
-        self,
-        *args,
-        feature_effect_scale=1.0,
-        unit_effect_scale=0.5,
-        time_effect_scale=0.3,
-        noise_scale=0.2,
-        treatment_effect_value=2.0,
-        confounding_strength=0.3,
-        **kwargs,
-    ):
-        """Generate synthetic panel data for synthetic control analysis.
-
-        Parameters
-        ----------
-        feature_effect_scale : float, default=1.0
-            Scale parameter for the effects of features on the outcome.
-            Feature weights are drawn from N(0, feature_effect_scale^2).
-        unit_effect_scale : float, default=0.5
-            Scale parameter for the random component of unit-specific effects.
-        time_effect_scale : float, default=0.3
-            Scale parameter for time-specific effects.
-        noise_scale : float, default=0.2
-            Scale parameter for random noise in the outcome.
-        treatment_effect_value : float, default=2.0
-            The magnitude of the treatment effect.
-        confounding_strength : float, default=0.3
-            Strength of confounding. Unit effects are generated as
-            `confounding_strength * features[:, 0] + N(0, unit_effect_scale^2)`.
-            If n_features is 0, this parameter has no effect.
-
-        Returns
-        -------
-        dict
-            Dictionary containing:
-            - 'df': pandas DataFrame with panel data.
-            - 'true_treatment_effect': The true treatment effect value (scalar).
-            - 'unit_effects': numpy array of unit-specific effects.
-            - 'time_effects': numpy array of time-specific effects.
-            - 'features': numpy array of features (n_total_units x n_features).
-            - 'feature_weights': numpy array of weights for features.
-            - 'n_treated_units': Number of treated units.
-            - 'n_control_units': Number of control units.
-            - 'n_time_pre': Number of pre-treatment periods.
-            - 'n_time_post': Number of post-treatment periods.
-        """
-        if args:
-            raise ValueError(f"Unexpected positional arguments: {args}")
-        if kwargs:
-            raise ValueError(f"Unexpected keyword arguments: {kwargs}")
-        self.set_seed()
-
-        n_total_units = self.n_treated_units + self.n_control_units
-        n_total_time = self.n_time_pre + self.n_time_post
-
-        unit_ids = np.arange(n_total_units)
-        time_ids = np.arange(n_total_time)
-
-        features = np.random.normal(size=(n_total_units, self.n_features))
-
-        confounded_part = np.zeros(n_total_units)
-        if self.n_features > 0 and confounding_strength != 0:
-            confounded_part = confounding_strength * features[:, 0]
-
-        random_part_unit_effects = np.random.normal(size=n_total_units, scale=unit_effect_scale)
-        unit_effects = confounded_part + random_part_unit_effects
-
-        time_effects_trend = time_effect_scale * np.arange(n_total_time)
-        time_effects_noise = np.random.normal(size=n_total_time, scale=time_effect_scale / 2)
-        time_effects = time_effects_trend + time_effects_noise
-
-        feature_weights = np.array([])
-        if self.n_features > 0:
-            feature_weights = np.random.normal(loc=0.0, scale=feature_effect_scale, size=self.n_features)
-
-        outcomes = np.zeros((n_total_units, n_total_time))
-        for i in range(n_total_units):
-            for t in range(n_total_time):
-                outcome_val = unit_effects[i] + time_effects[t]
-
-                if self.n_features > 0:
-                    outcome_val += np.dot(features[i, :], feature_weights)
-
-                is_treated_unit = i < self.n_treated_units
-                is_post_period = t >= self.n_time_pre
-                if is_treated_unit and is_post_period:
-                    outcome_val += treatment_effect_value
-
-                outcome_val += np.random.normal(scale=noise_scale)
-                outcomes[i, t] = outcome_val
-
-        data = []
-        for i in unit_ids:
-            for t in time_ids:
-                row = {
-                    "unit_id": i,
-                    "time_id": t,
-                    "outcome": outcomes[i, t],
-                    "is_treated_unit": i < self.n_treated_units,
-                    "is_post_period": t >= self.n_time_pre,
-                }
-                for j in range(self.n_features):
-                    row[f"X{j + 1}"] = features[i, j]
-                data.append(row)
-
-        df = pd.DataFrame(data)
-
-        return {
-            "df": df,
-            "true_treatment_effect": treatment_effect_value,
-            "unit_effects": unit_effects,
-            "time_effects": time_effects,
-            "features": features,
-            "feature_weights": feature_weights,
-            "n_treated_units": self.n_treated_units,
-            "n_control_units": self.n_control_units,
-            "n_time_pre": self.n_time_pre,
-            "n_time_post": self.n_time_post,
-        }
-
-
-class DRDiDSC(BaseDGP):
-    """Data Generating Process for Doubly Robust DiD Synthetic Control methods."""
-
-    def __init__(
-        self,
-        n_units: int = 100,
-        n_time_pre: int = 10,
-        n_time_post: int = 5,
-        n_features: int = 3,
-        n_latent_factors: int = 2,
-        random_seed: int = 42,
-    ):
-        """Initialize the DRDiDSC data generating process.
-
-        Parameters
-        ----------
-        n_units : int, default=100
-            Total number of units.
-        n_time_pre : int, default=10
-            Number of time periods before treatment.
-        n_time_post : int, default=5
-            Number of time periods after treatment starts.
-        n_features : int, default=3
-            Number of time-invariant covariates to generate.
-        n_latent_factors : int, default=2
-            Number of latent factors in the outcome model.
-        random_seed : int, default=42
-            Random seed for reproducibility.
-        """
-        super().__init__(random_seed=random_seed)
-        self.n_units = n_units
-        self.n_time_pre = n_time_pre
-        self.n_time_post = n_time_post
-        self.n_total_time = n_time_pre + n_time_post
-        self.treatment_time = n_time_pre
-        self.n_features = n_features
-        self.n_latent_factors = n_latent_factors
-
-    def generate_data(
-        self,
-        *args,
-        feature_mean: float | np.ndarray = 0.0,
-        feature_cov: np.ndarray | None = None,
-        prop_score_coefs: np.ndarray | None = None,
-        prop_score_intercept: float = 0.0,
-        treatment_fraction_target: float = 0.2,
-        factor_loadings_scale: float = 0.5,
-        time_factors_scale: float = 0.5,
-        outcome_coefs: np.ndarray | None = None,
-        unit_fe_scale: float = 0.3,
-        time_fe_scale: float = 0.2,
-        time_trend_strength: float = 0.05,
-        noise_scale: float = 0.1,
-        att_base: float = 1.0,
-        att_hetero_coefs: np.ndarray | None = None,
-        att_dynamic_coeffs: list[float] | np.ndarray | None = None,
-        **kwargs,
-    ):
-        """Generate synthetic panel data for DR-DiD-SC analysis.
-
-        Parameters
-        ----------
-        feature_mean : float or ndarray, default=0.0
-            Mean for covariate generation. Scalar or array of shape (n_features,).
-        feature_cov : ndarray or None, default=None
-            Covariance matrix for covariate generation, shape (n_features, n_features).
-            If None, defaults to identity matrix.
-        prop_score_coefs : ndarray or None, default=None
-            Coefficients for features in the propensity score model, shape (n_features,).
-            If None and n_features > 0, defaults to [1.0, 0, ...].
-        prop_score_intercept : float, default=0.0
-            Intercept for the propensity score model.
-        treatment_fraction_target : float, default=0.2
-            Target fraction of units to be treated.
-        factor_loadings_scale : float, default=0.5
-            Scale (std dev) for random factor loadings L_i.
-        time_factors_scale : float, default=0.5
-            Scale (std dev) for random time factors F_t.
-        outcome_coefs : ndarray or None, default=None
-            Coefficients for features in the Y(0) outcome model, shape (n_features,).
-            If None and n_features > 0, defaults to [0.5, 0, ...].
-        unit_fe_scale : float, default=0.3
-            Scale (std dev) for random unit fixed effects alpha_i.
-        time_fe_scale : float, default=0.2
-            Scale (std dev) for random component of time fixed effects delta_t.
-        time_trend_strength : float, default=0.05
-             Coefficient for a linear time trend component in delta_t.
-        noise_scale : float, default=0.1
-            Scale (std dev) for idiosyncratic error term epsilon_it.
-        att_base : float, default=1.0
-            Base average treatment effect on the treated.
-        att_hetero_coefs : ndarray or None, default=None
-            Coefficients for feature-based heterogeneity in ATT, shape (n_features,).
-            If None, no heterogeneity based on X.
-        att_dynamic_coeffs : list[float], ndarray or None, default=None
-            Coefficients for dynamic treatment effects relative to treatment_time.
-            e.g., [1.0, 1.2, 0.8] means effect is scaled by 1.0 in period T0, 1.2 in T0+1, etc.
-            If None, effect is constant as per att_base and att_hetero_coefs.
-
-        Returns
-        -------
-        dict
-            Dictionary containing detailed components of the DGP.
-        """
-        if args:
-            raise ValueError(f"Unexpected positional arguments: {args}")
-        if kwargs:
-            raise ValueError(f"Unexpected keyword arguments: {kwargs}")
-        self.set_seed()
-
-        _feature_mean_vec: np.ndarray
-        if isinstance(feature_mean, int | float):
-            _feature_mean_vec = np.full(self.n_features, float(feature_mean))
-        else:
-            _feature_mean_vec = np.asarray(feature_mean, dtype=float)
-            if _feature_mean_vec.shape != (self.n_features,):
-                raise ValueError("feature_mean must be scalar or (n_features,)")
-
-        _feature_cov_matrix: np.ndarray
-        if feature_cov is None:
-            _feature_cov_matrix = np.eye(self.n_features)
-        else:
-            _feature_cov_matrix = np.asarray(feature_cov, dtype=float)
-            if _feature_cov_matrix.shape != (self.n_features, self.n_features):
-                raise ValueError("feature_cov must be (n_features, n_features)")
-
-        X: np.ndarray
-        if self.n_features > 0:
-            X = np.random.multivariate_normal(_feature_mean_vec, _feature_cov_matrix, self.n_units)
-        else:
-            X = np.empty((self.n_units, 0))
-
-        _prop_score_coefs_actual = np.array([], dtype=float)
-        linear_propensity: np.ndarray
-        if self.n_features > 0:
-            if prop_score_coefs is None:
-                _prop_score_coefs_actual = np.zeros(self.n_features)
-                _prop_score_coefs_actual[0] = 1.0
-            else:
-                _prop_score_coefs_actual = np.asarray(prop_score_coefs, dtype=float)
-                if _prop_score_coefs_actual.shape != (self.n_features,):
-                    raise ValueError("prop_score_coefs must be (n_features,)")
-            linear_propensity = X @ _prop_score_coefs_actual + prop_score_intercept
-        else:
-            linear_propensity = np.full(self.n_units, prop_score_intercept, dtype=float)
-
-        true_prop_scores = 1 / (1 + np.exp(-linear_propensity))
-
-        D = np.zeros(self.n_units, dtype=bool)
-        if self.n_units > 0:
-            n_treated_target = int(self.n_units * treatment_fraction_target)
-            n_treated_actual = n_treated_target
-
-            if n_treated_target >= self.n_units and self.n_units > 1:
-                n_treated_actual = self.n_units - 1
-            elif n_treated_target >= self.n_units and self.n_units <= 1:
-                n_treated_actual = 0
-
-            n_treated_actual = max(n_treated_actual, 0)
-
-            if n_treated_actual > 0:
-                sorted_indices = np.argsort(true_prop_scores)
-                D[sorted_indices[-n_treated_actual:]] = True
-
-        L = np.random.normal(scale=factor_loadings_scale, size=(self.n_units, self.n_latent_factors))
-        F = np.random.normal(scale=time_factors_scale, size=(self.n_total_time, self.n_latent_factors))
-        factor_model_component = L @ F.T
-
-        _outcome_coefs_actual = np.array([], dtype=float)
-        direct_covariate_effect: np.ndarray
-        if self.n_features > 0:
-            if outcome_coefs is None:
-                _outcome_coefs_actual = np.zeros(self.n_features)
-                _outcome_coefs_actual[0] = 0.5
-            else:
-                _outcome_coefs_actual = np.asarray(outcome_coefs, dtype=float)
-                if _outcome_coefs_actual.shape != (self.n_features,):
-                    raise ValueError("outcome_coefs must be (n_features,)")
-            direct_covariate_effect = (X @ _outcome_coefs_actual)[:, np.newaxis]
-        else:
-            direct_covariate_effect = np.zeros((self.n_units, 1))
-
-        alpha_i = np.random.normal(scale=unit_fe_scale, size=self.n_units)[:, np.newaxis]
-
-        time_rng = np.arange(self.n_total_time)
-        delta_t_random_component = np.random.normal(scale=time_fe_scale, size=self.n_total_time)
-        delta_t_trend_component = time_trend_strength * time_rng
-        delta_t = (delta_t_random_component + delta_t_trend_component)[np.newaxis, :]
-
-        epsilon_it = np.random.normal(scale=noise_scale, size=(self.n_units, self.n_total_time))
-        Y0_it = direct_covariate_effect + factor_model_component + alpha_i + delta_t + epsilon_it
-
-        _att_hetero_coefs_actual = np.array([], dtype=float)
-        hetero_component_for_tau = np.zeros((self.n_units, 1))
-        if self.n_features > 0 and att_hetero_coefs is not None:
-            _att_hetero_coefs_actual = np.asarray(att_hetero_coefs, dtype=float)
-            if _att_hetero_coefs_actual.shape != (self.n_features,):
-                raise ValueError("att_hetero_coefs must be (n_features,)")
-            hetero_component_for_tau = (X @ _att_hetero_coefs_actual)[:, np.newaxis]
-
-        unit_base_att = att_base + hetero_component_for_tau
-
-        actual_tau_it = np.zeros((self.n_units, self.n_total_time))
-        _att_dynamic_coeffs_actual: np.ndarray
-        if att_dynamic_coeffs is None:
-            _att_dynamic_coeffs_actual = np.array([1.0])
-        else:
-            _att_dynamic_coeffs_actual = np.asarray(att_dynamic_coeffs, dtype=float)
-
-        for i in range(self.n_units):
-            if D[i]:
-                for t in range(self.treatment_time, self.n_total_time):
-                    time_since_treatment = t - self.treatment_time
-                    dyn_coeff_idx = min(time_since_treatment, len(_att_dynamic_coeffs_actual) - 1)
-                    dynamic_multiplier = _att_dynamic_coeffs_actual[dyn_coeff_idx]
-                    actual_tau_it[i, t] = unit_base_att[i, 0] * dynamic_multiplier
-
-        Y_it = Y0_it + actual_tau_it
-
-        unit_ids_col = np.repeat(np.arange(self.n_units), self.n_total_time)
-        time_ids_col = np.tile(np.arange(self.n_total_time), self.n_units)
-
-        df_data = {
-            "unit_id": unit_ids_col,
-            "time_id": time_ids_col,
-            "outcome": Y_it.ravel(),
-            "is_treated_unit": np.repeat(D, self.n_total_time),
-        }
-
-        treatment_active_in_period = np.zeros((self.n_units, self.n_total_time), dtype=bool)
-        for i in range(self.n_units):
-            if D[i]:
-                treatment_active_in_period[i, self.treatment_time :] = True
-        df_data["treatment_in_period"] = treatment_active_in_period.ravel()
-
-        for j in range(self.n_features):
-            df_data[f"X{j + 1}"] = np.repeat(X[:, j], self.n_total_time)
-        df = pd.DataFrame(df_data)
-
-        true_att_values = actual_tau_it[D, self.treatment_time :]
-        avg_true_att = np.mean(true_att_values) if true_att_values.size > 0 else np.nan
-
-        return {
-            "df": df,
-            "X": X,
-            "Y0_it": Y0_it,
-            "actual_tau_it": actual_tau_it,
-            "true_prop_scores": true_prop_scores,
-            "D_treatment_assignment": D,
-            "L_factor_loadings": L,
-            "F_time_factors": F,
-            "alpha_i_unit_effects": alpha_i.flatten(),
-            "delta_t_time_effects": delta_t.flatten(),
-            "beta_prop_propensity_coefs": _prop_score_coefs_actual,
-            "beta_outcome_outcome_coefs": _outcome_coefs_actual,
-            "att_hetero_coefs": _att_hetero_coefs_actual,
-            "avg_true_att": avg_true_att,
-            "n_treated_units_actual": np.sum(D),
-            "n_control_units_actual": self.n_units - np.sum(D),
-            "treatment_time": self.treatment_time,
-        }
-
-
 class SantAnnaZhaoDRDiD(BaseDGP):
     """Data Generating Process for Sant'Anna and Zhao (2020)'s DGP 1 (RC Case)."""
