+++ tests/test_utils.py
@@ -7,8 +7,8 @@ from .helpers import importorskip
 
 pd = importorskip("pandas")
 
-from pydid.utils import preprocess_drdid, preprocess_synth
-from tests.dgp import DiD, SantAnnaZhaoDRDiD, SyntheticControl
+from pydid.utils import preprocess_drdid
+from tests.dgp import DiD, SantAnnaZhaoDRDiD
 
 
 def test_pre_process_panel():
@@ -397,403 +397,4 @@ def test_preprocess_drdid_panel_duplicate_id_time():
             id_col="id",
             treat_col="is_treated_unit",
             panel=True,
-        )
-
-
-def test_preprocess_synth_basic():
-    dgp_params = {
-        "n_treated_units": 2,
-        "n_control_units": 10,
-        "n_time_pre": 5,
-        "n_time_post": 3,
-        "n_features": 2,
-    }
-    dgp = SyntheticControl(**dgp_params)
-    data_dict = dgp.generate_data()
-    df = data_dict["df"]
-
-    y_col = "outcome"
-    time_col = "time_id"
-    id_col = "unit_id"
-    treat_col = "is_treated_unit"
-    treatment_period = dgp_params["n_time_pre"]
-    l_outcome_periods = [treatment_period - 2, treatment_period - 1]
-    post_periods_of_interest = [treatment_period, treatment_period + 1]
-
-    result = preprocess_synth(
-        data=df,
-        y_col=y_col,
-        time_col=time_col,
-        id_col=id_col,
-        treat_col=treat_col,
-        treatment_period=treatment_period,
-        covariates_formula="~ X1 + X2",
-        l_outcome_periods=l_outcome_periods,
-        post_periods_of_interest=post_periods_of_interest,
-        normalized=True,
-    )
-
-    assert isinstance(result, dict)
-
-    expected_keys = [
-        "Y_pre_treat",
-        "Y_post_treat",
-        "Y_pre_control",
-        "Y_post_control",
-        "Z_treat",
-        "Z_control",
-        "weights_treat",
-        "weights_control",
-        "n_treated",
-        "n_control",
-        "n_pre_periods",
-        "n_post_periods",
-        "pre_periods",
-        "post_periods_of_interest",
-        "treatment_period",
-        "predictor_names",
-        "treated_ids",
-        "control_ids",
-        "normalized_weights",
-    ]
-    for key in expected_keys:
-        assert key in result, f"Key '{key}' missing from result."
-
-    assert result["n_treated"] == dgp_params["n_treated_units"]
-    assert result["n_control"] == dgp_params["n_control_units"]
-    assert result["n_pre_periods"] == dgp_params["n_time_pre"]
-    assert result["n_post_periods"] == len(post_periods_of_interest)
-    assert result["Y_pre_treat"].shape == (
-        dgp_params["n_time_pre"],
-        dgp_params["n_treated_units"],
-    )
-    assert result["Y_post_treat"].shape == (
-        len(post_periods_of_interest),
-        dgp_params["n_treated_units"],
-    )
-    assert result["Y_pre_control"].shape == (
-        dgp_params["n_time_pre"],
-        dgp_params["n_control_units"],
-    )
-    assert result["Y_post_control"].shape == (
-        len(post_periods_of_interest),
-        dgp_params["n_control_units"],
-    )
-
-    num_lagged_outcomes = len(l_outcome_periods)
-    num_covariates = dgp_params["n_features"]
-    expected_num_predictors = num_lagged_outcomes + num_covariates
-
-    assert len(result["predictor_names"]) == expected_num_predictors
-    if expected_num_predictors > 0:
-        assert result["Z_treat"].shape == (
-            expected_num_predictors,
-            dgp_params["n_treated_units"],
-        )
-        assert result["Z_control"].shape == (
-            expected_num_predictors,
-            dgp_params["n_control_units"],
-        )
-    else:
-        assert result["Z_treat"].shape[0] == 0
-        assert result["Z_control"].shape[0] == 0
-
-    assert len(result["weights_treat"]) == dgp_params["n_treated_units"]
-    assert len(result["weights_control"]) == dgp_params["n_control_units"]
-    assert result["normalized_weights"] is True
-    assert result["treatment_period"] == treatment_period
-    assert len(result["pre_periods"]) == dgp_params["n_time_pre"]
-    assert result["post_periods_of_interest"] == post_periods_of_interest
-    assert len(result["treated_ids"]) == dgp_params["n_treated_units"]
-    assert len(result["control_ids"]) == dgp_params["n_control_units"]
-
-
-def test_preprocess_synth_empty_df():
-    empty_df = pd.DataFrame()
-    with pytest.raises(ValueError, match="Missing required columns"):
-        preprocess_synth(
-            data=empty_df,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=1,
-        )
-
-
-def test_preprocess_synth_treatment_period_not_in_time():
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 0, 1],
-            "outcome": [1, 2, 3, 4],
-            "id": [1, 1, 2, 2],
-            "treatment": [0, 0, 1, 1],
-        }
-    )
-    with pytest.raises(ValueError, match="treatment_period .* not found in time column"):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=2,
-        )
-
-
-def test_preprocess_synth_treatment_period_is_first():
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 2, 0, 1, 2],
-            "outcome": [1, 2, 3, 4, 5, 6],
-            "id": [1, 1, 1, 2, 2, 2],
-            "treatment": [0, 0, 0, 1, 1, 1],
-        }
-    )
-    with pytest.raises(ValueError, match="treatment_period cannot be the first period"):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=0,
-        )
-
-
-def test_preprocess_synth_no_pre_treatment_periods():
-    data = pd.DataFrame(
-        {
-            "time": [1, 2, 3, 1, 2, 3],
-            "outcome": [1, 2, 3, 4, 5, 6],
-            "id": [1, 1, 1, 2, 2, 2],
-            "treatment": [0, 0, 0, 1, 1, 1],
-        }
-    )
-    with pytest.raises(ValueError, match="treatment_period cannot be the first period"):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=1,
-        )
-
-
-def test_preprocess_synth_non_numeric_weights():
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 2, 0, 1, 2, 0, 1, 2],
-            "outcome": [1, 2, 3, 4, 5, 6, 7, 8, 9],
-            "id": [1, 1, 1, 2, 2, 2, 3, 3, 3],
-            "treatment": [0, 0, 0, 1, 1, 1, 0, 0, 0],
-            "w": ["a", "a", "a", "b", "b", "b", "c", "c", "c"],
-        }
-    )
-    with pytest.raises(TypeError, match="Weights column 'w' must be numeric"):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=1,
-            weights_col="w",
-        )
-
-
-def test_preprocess_synth_zero_weights():
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 2, 0, 1, 2, 0, 1, 2],
-            "outcome": [1, 2, 3, 4, 5, 6, 7, 8, 9],
-            "id": [1, 1, 1, 2, 2, 2, 3, 3, 3],
-            "treatment": [0, 0, 0, 1, 1, 1, 0, 0, 0],
-            "w": [0, 0, 0, 0, 0, 0, 0, 0, 0],
-        }
-    )
-    with pytest.warns(UserWarning, match="Mean of weights is zero or negative. Cannot normalize."):
-        result = preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=1,
-            weights_col="w",
-            normalized=True,
-        )
-    assert "weights_treat" in result
-    assert "weights_control" in result
-    assert np.all(result["weights_treat"] == 0)
-    assert np.all(result["weights_control"] == 0)
-
-
-def test_preprocess_synth_collinearity_warning():
-    data = pd.DataFrame(
-        {
-            "time": [0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3],
-            "id": [1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4],
-            "outcome": [5, 6, 7, 20, 10, 11, 12, 21, 13, 14, 15, 22, 16, 17, 18, 23],
-            "X1": [0, 0, 0, 0, 10, 11, 12, 21, 0, 0, 0, 0, 0, 0, 0, 0],
-            "treatment": [0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0],
-        }
-    )
-    data["is_treated_unit"] = data["id"].map(data.groupby("id")["treatment"].first())
-
-    with pytest.warns(UserWarning, match="Potential collinearity detected among predictors"):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="is_treated_unit",
-            treatment_period=2,
-            l_outcome_periods=[1],
-            covariates_formula="~ X1",
-        )
-
-
-def test_preprocess_synth_negative_weights():
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 2, 0, 1, 2, 0, 1, 2],
-            "outcome": [1, 2, 3, 4, 5, 6, 7, 8, 9],
-            "id": [1, 1, 1, 2, 2, 2, 3, 3, 3],
-            "treatment": [0, 0, 0, 1, 1, 1, 0, 0, 0],
-            "w": [1, 1, 1, -1, -1, -1, 1, 1, 1],
-        }
-    )
-    with pytest.warns(UserWarning, match="Some weights are negative."):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=1,
-            weights_col="w",
-        )
-
-
-@pytest.mark.parametrize("scenario", ["no_treated_after_balance", "no_control_after_balance"])
-def test_preprocess_synth_no_units_after_balancing(scenario):
-    base_data = {
-        "time": [0, 1, 2, 3, 0, 1, 2, 3, 0, 1, 2, 3],
-        "outcome": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
-        "id": [1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3],
-        "treatment": [0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 0],
-    }
-    df = pd.DataFrame(base_data)
-
-    treatment_period = 2
-
-    if scenario == "no_treated_after_balance":
-        df_modified = df[~((df["id"] == 2) & (df["time"] == 0))]
-        expected_error_match = "No treated units remain after balancing panel"
-    else:
-        df_modified = df[~((df["id"] == 1) & (df["time"] == 1))]
-        df_modified = df_modified[~((df_modified["id"] == 3) & (df_modified["time"] == 0))]
-        expected_error_match = "No control units remain after balancing panel"
-
-    with pytest.raises(ValueError, match=expected_error_match):
-        preprocess_synth(
-            data=df_modified,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=treatment_period,
-            post_periods_of_interest=[2, 3],
-        )
-
-
-def test_preprocess_synth_invalid_l_outcome_periods():
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 2, 3, 0, 1, 2, 3],
-            "outcome": [1, 2, 3, 4, 5, 6, 7, 8],
-            "id": [1, 1, 1, 1, 2, 2, 2, 2],
-            "treatment": [0, 0, 0, 0, 1, 1, 1, 1],
-        }
-    )
-    with pytest.raises(ValueError, match="All l_outcome_periods must be in the pre-treatment period"):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=2,
-            l_outcome_periods=[0, 1, 2],
-        )
-
-
-@pytest.mark.parametrize("invalid_post_periods", [([0, 1, 2]), ([2, 3, 4])])
-def test_preprocess_synth_invalid_post_periods_of_interest(invalid_post_periods):
-    data = pd.DataFrame(
-        {
-            "time": [0, 1, 2, 3, 0, 1, 2, 3],
-            "outcome": [1, 2, 3, 4, 5, 6, 7, 8],
-            "id": [1, 1, 1, 1, 2, 2, 2, 2],
-            "treatment": [0, 0, 0, 0, 1, 1, 1, 1],
-        }
-    )
-    match_str = (
-        "All post_periods_of_interest must be >= treatment_period"
-        if any(p < 2 for p in invalid_post_periods)
-        else "Not all post_periods_of_interest found in time column"
-    )
-    with pytest.raises(ValueError, match=match_str):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=2,
-            post_periods_of_interest=invalid_post_periods,
-        )
-
-
-@pytest.mark.parametrize(
-    "treatment_setup",
-    [
-        {
-            "data": pd.DataFrame(
-                {
-                    "time": [0, 1, 0, 1],
-                    "outcome": [1, 2, 3, 4],
-                    "id": [1, 1, 2, 2],
-                    "treatment": [0, 0, 0, 0],
-                }
-            ),
-            "match": "No treated units found",
-        },
-        {
-            "data": pd.DataFrame(
-                {
-                    "time": [0, 1, 0, 1],
-                    "outcome": [1, 2, 3, 4],
-                    "id": [1, 1, 2, 2],
-                    "treatment": [1, 1, 1, 1],
-                }
-            ),
-            "match": "No control units found",
-        },
-    ],
-)
-def test_preprocess_synth_no_treated_or_control_units(treatment_setup):
-    data = treatment_setup["data"]
-    match_error = treatment_setup["match"]
-
-    with pytest.raises(ValueError, match=match_error):
-        preprocess_synth(
-            data=data,
-            y_col="outcome",
-            time_col="time",
-            id_col="id",
-            treat_col="treatment",
-            treatment_period=1,
         )
